requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.1.1
pydantic[email]==2.11.7
qrcode[pil]==7.4.2
Pillow==11.0.0
alembic==1.16.4
PyJWT==2.8.0
